Vivaldi 1.0.403.24 (Beta 3)
Vivaldi getting closer to 1.0. Beta 3 is here with New session load/save, tab zoom and a lot of new options!


Dear Friends,

Spring is just around the corner so is Vivaldi 1.0. While we are waiting for it, we thought we would share the latest update with you.

Since the Beta 2, our primary focus for this release has been on improving performance and stability, as well as polishing the design throughout Vivaldi UI. This includes adding a lot of new settings as always :) and we have also added a few new features.

 

Here are some of the highlights for the new features and changes in this Beta 3 release:

A sessions save/load
This is our first implementation of the session manager. In this implementation, you can save all of your currently open tabs as a session so that you can go back and open them all at once. So for example, you can have a work session different from a private session. That way, you can very quickly switch between completely different set of tabs for work, and for your private research or what not.

To save a session, go to File menu and click “Save Open Tabs as Session”.



 

And to open a session, Go to Open a session under the File menu and choose from the saved sessions. 






Tab specific zoom allows you to set zoom level for each tab, rather than for each website domain. It avoids having all the tabs zoomed when visiting the same website in different tabs and allows you to keep one zoom level while browsing in a tab.



 

New Tab Hibernation Options
This new option gives you more control on your resources. If you are like us and want to have many tabs open, simply right click on the active tab and choose 'Hibernation Background tabs'. This will immediately free up more resources as your tabs are 'put to sleep'. Switching over to any of them will bring them back to life.

.

 

New options and defaults for tab opening and closing behaviour
Another improvement from Beta 2 based on user feedback is tab opening/closing behaviour. When you open new tabs from a link in your current tab, they share a relationship. For example, say you perform a search in your current tab and Ctrl+Click the first search result, it will then open next to its "parent". If you Ctrl+Click a second result, the new tab will place itself immediately after its "sibling".

New tabs that you open via Ctrl+T or by clicking on "+" have no relationships and will therefore open on next to the "+" button, just like before.

Opening behaviour goes hand in hand with closing behaviour, so we have also made a change here as well. With the new default, when a focused tab is closed, Vivaldi will switch to its right-most (or bottom-most) "relative". In cases where there are no relationships between tabs, Vivaldi will switch to the last activated tab (our old behavior). This combination of tab relationships and activation order, should result in a best of both worlds default and an improvement over what the other browsers are doing.

The new default works especially well when opening lots of links from one site and then closing them down one by one to return to whatever you were working on before.

But of course, all of this can be customized as you like from Tabs menu in the Settings.



 

A suggestion for our regular snapshot users, if you’ve downloaded our recent snapshot, a new installation is not required as this includes the same fixes that you received in the last snapshot.

Thank you always to all of you for supporting us and active contributor to our development as well as being strong advocate for Vivaldi. Enjoy the new beta!

 
Published in: Vivaldi updates
Written by Tatsuki Tomita
March 4, 2016